change logic for update restaurants
