+++ holiholic_planner/src/main/java/com/holiholic/planner/HoliholicPlannerApplication.java
@@ -1,6 +1,7 @@
 package com.holiholic.planner;
 
 import com.holiholic.planner.database.DatabaseManager;
+import com.holiholic.planner.update.parser.ParserManager;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
@@ -10,5 +11,6 @@ public class HoliholicPlannerApplication {
 	public static void main(String[] args) {
 		SpringApplication.run(HoliholicPlannerApplication.class, args);
         DatabaseManager.setLogger();
+		ParserManager.setLogger();
 	}
 }
