+++ holiholic_planner/src/main/java/com/holiholic/planner/update/action/Action.java
@@ -1,0 +1,8 @@
+package com.holiholic.planner.update.action;
+
+ /* Action - Interface to implement for future commands
+ *
+ */
+public interface Action {
+    boolean execute();
+}
